Correct misnamed base component in compose file
The slack API export base image was misnamed in the docker compose file,
leading builds to fail in certain circumstances. Other components that
depended on the base image would fail to build if the base image hadn't
previously been built on the system. If it was not present, the docker
compose would build it, but was using the incorrect name, leading to the
failed builds when the dependent images were being built.